requirements.txt
django>=5.0.1,<6.0.0 # 5.0.1 이상 6.0.0 미만
djangorestframework>=3.14.0,<4.0.0


psycopg2-binary>=2.9.9,<3.0
# psycopg2 설치 (django postgresql)
# psycopg2>=2.9.9,<3.0

#drf-spectacular
drf-spectacular>=0.27.1,<0.28.0


# Help Djang Chatting Socket Connection
channels>=4.0.0,<4.0.1